[vulkan] Move Tensor.* from ops/ to api/
Moves `Tensor.h` and `Tensor.cpp` from the `ops/` folder to the `api/` folder.

Differential Revision: [D42103700](https://our.internmc.facebook.com/intern/diff/D42103700/)

[ghstack-poisoned]
